worker: split download.go and remove debug code
